requirements.txt
pyyaml
python-telegram-bot
telethon
